refactor: 쿠키 생성 책임 위치 수정 및 불필요한 인증 책임 분리 수정
