+++ src/main/java/roomescape/service/AuthenticationService.java
@@ -1,27 +1,42 @@
 package roomescape.service;
 
-import jakarta.servlet.http.Cookie;
+import io.jsonwebtoken.*;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import roomescape.domain.AuthenticationGenerator;
+
+import java.util.Date;
 
 @Service
 public class AuthenticationService {
 
-    private final AuthenticationGenerator authenticationGenerator;
-
-    public AuthenticationService(AuthenticationGenerator authenticationGenerator) {
-        this.authenticationGenerator = authenticationGenerator;
-    }
+    @Value("${security.jwt.token.secret-key}")
+    private String secretKey;
+    @Value("${security.jwt.token.expire-length}")
+    private long validityInMilliseconds;
 
     public String createToken(String email) {
-        return authenticationGenerator.createToken(email);
-    }
-
-    public Cookie createCookie(String token) {
-        return authenticationGenerator.createCookie(token);
+        return Jwts.builder()
+                .setSubject(email)
+                .claim("email", email)
+                .setExpiration(new Date(System.currentTimeMillis() + validityInMilliseconds))
+                .signWith(SignatureAlgorithm.HS256, secretKey)
+                .compact();
     }
 
     public String getPayload(String token) {
-        return authenticationGenerator.getPayload(token);
+        if (validateToken(token)) {
+            return Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token).getBody().getSubject();
+        }
+        throw new IllegalStateException("올바른 토큰이 아닙니다.");
+    }
+
+    public boolean validateToken(String token) {
+        try {
+            Jws<Claims> claims = Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token);
+
+            return !claims.getBody().getExpiration().before(new Date());
+        } catch (JwtException | IllegalArgumentException e) {
+            return false;
+        }
     }
 }
